// src/app/master/model/department.ts
export class DepartmentView {
    name: string = "";
    id: string = "";
    sno: number = 0;
    status: boolean = false;
}

export class Departmentupdate {
    name: string = "";
    id: string = "";
    status: boolean = false;
}

export class Department {
    name: string = "";
    id: string = "";
    status: boolean = false;
}